requirements.txt
#AI Frameworks
langchain
llama-index
crewai
chromadb

# LLMs & Embeddings
openai
transformers
sentence-transformers

# Backend & API
fastapi
uvicorn
pydantic

# Database Drivers
sqlalchemy
psycopg2-binary #PostgreSQL
pyodbc #SQL Server
mysql-connector-python
snowflake-connector-python

#Additional Tools
tqdm
python-dotenv #Environment variable
loguru #Logging

# Testing
pytest

#pip install sqlparse
